pyproject.toml
[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[project]
name = "pdfquery"
version = "0.1.0"
description = "CLI to index PDFs into FAISS and query them with GPT"
authors = [{ name = "Steve Ghani", email = "maazghani@gmail.com" }]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "openai>=1.14.2",
    "faiss-cpu>=1.7.4",
    "PyPDF2>=3.0.1",
    "numpy>=1.24",
    "typer[all]>=0.9",
    "tqdm>=4.66",
]

[project.scripts]
pdfquery = "pdfquery.cli:main"
